Relocate Pipeline orchestrators
